requirements.txt
streamlit
requests
opencv-python
numpy
inference-sdk
gTTS
pygame
google-generativeai
